+++ src/components/EditPreview/EditPreview.tsx
@@ -12,7 +12,16 @@ import './EditPreview.scss';
 import { useInputsState, useUIState } from '@src/store';
 
 export const EditPreview = memo(() => {
-  const { record, previewContent, setPreviewContent, resetPreviewContent } = useInputsState();
+  const {
+    record,
+    previewContent,
+    setPreviewContent,
+    resetPreviewContent,
+    resetRecord,
+    resetSelectedEntries,
+    resetUserValues,
+    resetSelectedRecordBlocks,
+  } = useInputsState();
   const { currentlyPreviewedEntityBfid } = useUIState();
   const isCreatePageOpen = useRoutePathPattern(RESOURCE_CREATE_URLS);
   const { resourceId } = useParams();
@@ -28,6 +37,18 @@ export const EditPreview = memo(() => {
   useEffect(() => {
     resetPreviewContent();
   }, [resourceId]);
+
+  // TODO: UILD-552 - Temporary solution to reset the preview content when the user opens the create work page.
+  // The store should be cleared when the user navigates outside the LDE.
+  useEffect(() => {
+    if (isCreateWorkPageOpened) {
+      resetPreviewContent();
+      resetRecord();
+      resetSelectedEntries();
+      resetUserValues();
+      resetSelectedRecordBlocks();
+    }
+  }, [isCreateWorkPageOpened]);
 
   return (
     <div