+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change history for ui-linked-data
 
 ## 1.0.6 (IN PROGRESS)
+* Fixed incorrect instances list display when the user opens the Create Work page. Fixes [UILD-560].
+
+[UILD-560]:https://folio-org.atlassian.net/browse/UILD-560
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].